requirements.txt
python-telegram-bot[callback-data]==20.6
python-dotenv
